requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.6
django-debug-toolbar==4.2.0
django-debug-toolbar-force==0.2
django-extensions==3.2.3
django-filter==23.3
django-nine==0.2.7
djangorestframework==3.14.0
executing==2.0.0
httpie==3.2.2
idna==3.4
ipython==8.16.1
jedi==0.19.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg==3.1.12
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
